Fix: 후원하기 토스트 바로 사라지는 버그 수정
